[Move `@kbn/config-schema` to server] presentation plugins
