+++ src/plugins/input_control_vis/server/config.ts
