+++ src/plugins/input_control_vis/server/index.ts
@@ -9,7 +9,7 @@
 import type { CoreSetup, PluginConfigDescriptor, PluginInitializerContext } from '@kbn/core/server';
 import type { VisualizationsServerSetup } from '@kbn/visualizations-plugin/server';
 
-import { configSchema, InputControlConfig } from '../config';
+import { configSchema, InputControlConfig } from './config';
 
 export const config: PluginConfigDescriptor<InputControlConfig> = {
   exposeToBrowser: {
